Add files via upload
The Paint Calculator app is an essential Android application designed for homeowners, DIY enthusiasts, and professional painters. The app helps users accurately determine the amount of paint needed to cover a specific area, ensuring efficient paint usage and cost management. Users simply input the dimensions of the area they wish to paint, and the app calculates the required amount of paint in gallons.